RDC fixes

+++ vmbuild/genrdc.ps1
@@ -1,21 +1,42 @@
 . $PSScriptRoot\Common.ps1
 $userName = "admin"
+$DesktopPath = [Environment]::GetFolderPath("Desktop")
+$savepath = Join-Path $DesktopPath "memlabs.rdg"
+$templatefile = Join-Path $PSScriptRoot "template.rdg"
+
 $result = Test-Configuration -FilePath "E:\repos\memlabs\vmbuild\config\samples\Standalone.json"
 #$result = Test-Configuration -FilePath "E:\repos\memlabs\vmbuild\config\samples\Hierarchy.json"
+#$result = Test-Configuration -FilePath "E:\repos\memlabs\vmbuild\test.json"
 #$result = Test-Configuration -FilePath "E:\repos\memlabs\vmbuild\config\samples\AddToExisting.json"
+
+
 if ($result.Valid) {
     $deployConfig = $result.DeployConfig
+}else{
+    Write-Host "Config file is not valid"
+    return
 }
 
 function New-RDCManFile {
     param(
         [string]$rdcmanfile
     )
-    $template = Get-RDCManTemplate $rdcmanfile
-
+    $template = Get-RDCManTemplate $templatefile
+    if ($null -eq $template){
+        Write-Host "Could not locate $templatefile"
+        return
+    }
+    $existing = Get-RDCManTemplateorExisting $rdcmanfile
+    if ($null -eq $template){
+        Write-Host "Could not locate $rdcmanfile"
+        return
+    }
     #This is the bulk of the data.
-    $file = $template.RDCMan.file
+    $file = $existing.RDCMan.file
     $group = $file.group
+
+    $groupFromTemplate = $template.RDCMan.file.group
+
     # <RDCMan>
     #   <file>
     #     <group>
@@ -25,13 +46,17 @@ function New-RDCManFile {
     #     <group>
     #     ...
 
-    $findGroup = Get-RDCManGroupToModify $group $findGroup
+    $findGroup = Get-RDCManGroupToModify $group $findGroup $groupFromTemplate $existing
     if ($findGroup -eq $false -or $null -eq $findGroup) {
         Write-Error "Error in Get-RDCManPassword"
         return
     }
+    $shouldSave = $False
     foreach ($vm in $deployConfig.virtualMachines) {
-        Add-RDCManServerToGroup $vm.vmName $group $findgroup
+        if (Add-RDCManServerToGroup $vm.vmName $group $findgroup $groupFromTemplate $existing -eq $True)
+        {
+            $shouldSave = $True
+        }
     }
 
     #If the original file was a template, remove the templated group.
@@ -41,10 +66,14 @@ function New-RDCManFile {
     [void]$file.AppendChild($findgroup)
 
     #Save to desired filename
-    Get-Process -Name rdcman -ea Ignore | Stop-Process
-    Start-Sleep 1
-    Write-Host "Saving resultant XML to $rdcmanfile"
-    $template.save($rdcmanfile) | Out-Null
+    if ($shouldSave -eq $True){
+        Get-Process -Name rdcman -ea Ignore | Stop-Process
+        Start-Sleep 1
+        Write-Host "Saving resultant XML to $rdcmanfile"
+        $existing.save($rdcmanfile) | Out-Null
+    } else{
+        Write-Host "No Changes. Not Saving resultant XML to $rdcmanfile"
+    }
 
 }
 
@@ -53,21 +82,27 @@ function Add-RDCManServerToGroup {
     param(
         [string]$serverName,
         $group,
-        $findgroup
+        $findgroup,
+        $groupFromTemplate,
+        $existing
     )
 
     $findserver = $findgroup.server | Where-Object { $_.properties.name -eq $serverName } | Select-Object -First 1
     Write-Host "Adding $serverName to RDG Group ... " -NoNewline
     if ($null -eq $findserver) {
         Write-Host "added"
-        $server = $group.SelectNodes('//server') | Select-Object -First 1
+        $server = $groupFromTemplate.SelectNodes('//server') | Select-Object -First 1
         $newserver = $server.clone()
-        $newserver.properties.name = $serverName
-        $findgroup.AppendChild($newserver)   | Out-Null
+        $newserver.properties.name = $serverName 
+        $clonedNode = $existing.ImportNode($newserver, $true)
+        $findgroup.AppendChild($clonedNode) 
+        return $True       
     }
     else {
         Write-Host "already exists in group. Skipping"
+        return $False
     }
+    return $False
 }
 
 #This gets the <Group> section from the template. Either makes a new one, or returns an existing one.
@@ -75,7 +110,7 @@ function Add-RDCManServerToGroup {
 function Get-RDCManGroupToModify {
     param(
         $group,
-        $findgroup
+        $groupFromTemplate
     )
     $domain = $deployConfig.vmOptions.domainName
     Write-Host "Looking for group entry named $domain in current xml... " -NoNewline
@@ -83,7 +118,7 @@ function Get-RDCManGroupToModify {
 
     if ($null -eq $findGroup) {
         Write-Host "Not found.  Creating new group"
-        $findGroup = $group.Clone()
+        $findGroup = $groupFromTemplate.Clone()
         $findGroup.properties.name = $domain
         $findGroup.logonCredentials.userName = $userName
         $password = Get-RDCManPassword
@@ -97,6 +132,7 @@ function Get-RDCManGroupToModify {
         foreach ($Child in $ChildNodes) {
             [void]$Child.ParentNode.RemoveChild($Child)
         }
+        $findGroup = $existing.ImportNode($findGroup, $true)
     }
     else {
         Write-Host "Found!"
@@ -107,16 +143,26 @@ function Get-RDCManGroupToModify {
 #Gets the blank template, or returns the existing rdg xml if available.
 function Get-RDCManTemplate {
     param(
-        [string]$rdcmanfile
+        [string]$templatefile
     )
-    $templatefile = Join-Path $PSScriptRoot "template.rdg"
-
-    if ((test-path "$rdcmanfile")) {
-        $templatefile = $rdcmanfile
-    }
-    Write-Host "Loading existing XML from $templatefile"
+    Write-Host "Loading template XML from $templatefile"
     [xml]$template = Get-Content -Path $templatefile
     return $template
+}
+
+#Gets the blank template, or returns the existing rdg xml if available.
+function Get-RDCManTemplateorExisting {
+    param(
+        [string]$rdcmanfile
+    )
+    $existingfile = Join-Path $PSScriptRoot "template.rdg"
+
+    if ((test-path "$rdcmanfile")) {
+        $existingfile = $rdcmanfile
+    }
+    Write-Host "Loading existing XML from $existingfile"
+    [xml]$existing = Get-Content -Path $existingfile
+    return $existing
 }
 
 function Get-RDCManPassword() {
@@ -137,6 +183,5 @@ function Get-RDCManPassword() {
     $EncryptionSettings = New-Object -TypeName RdcMan.EncryptionSettings
     return [RdcMan.Encryption]::EncryptString($Common.LocalAdmin.GetNetworkCredential().Password , $EncryptionSettings)
 }
-$DesktopPath = [Environment]::GetFolderPath("Desktop")
-$savepath = Join-Path $DesktopPath "memlabs.rdg"
+
 New-RDCManFile $savepath
